+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Changelog
 * :bug:`10158` PnL reports will now properly handle events with the same timestamp.
 * :bug:`10146` Users will now be able to filter assets in the history events filter if a non-EVM location is selected.
 * :bug:`10150` The history events page will be refreshed automatically when an asset is ignored or marked as spam.
+* :bug:`10199` Independentreserve trades, deposits and withdrawals now use the execution timestamp and not the creation one.
 
 * :release:`1.39.0 <2025-06-13>`
 * :bug:`-` Users will now be able to perform on-chain send transactions using older wallets such as Trust Wallet.